+++ test/test_sim_dmgp_wb.jl
@@ -109,4 +109,17 @@
         parv1[end-2] *= 2
         @test calc_lnlike(model, toas, parv1) < calc_lnlike(model, toas, params)
     end
+
+    @testset "posterior" begin
+        params = model.param_handler._default_params_tuple
+        parv = read_param_values_to_vector(model.param_handler, params)
+
+        calc_lnpost_ = get_lnpost_func(model, toas)
+        @test isfinite(calc_lnpost_(params))
+
+        calc_lnpost_vec = get_lnpost_func(model, toas, true)
+        paramss = transpose([parv parv parv])
+        @test allequal(calc_lnpost_vec(paramss))
+        @test calc_lnpost_vec(paramss)[1] ≈ calc_lnpost_(params)
+    end
 end